mds: support constness in MDSCacheObjects
So that one can have const CInode and CDir references
from time to time.

This is mostly just dumb adding 'const' to declarations
and iterators, but there is also an awkward case in CInode
where CINode::get_caps_issued was resetting loner_cap to -1
based on get_auth(), so that has changed functionally to
update loner_cap in state_set/state_clear/state_reset
instead so that the getter is just a getter.

Signed-off-by: John Spray <john.spray@redhat.com>